Moved around api routex
