+++ src/api/user/post_user.js
@@ -1,5 +1,5 @@
 import { collection, query, where, getDocs, addDoc } from 'firebase/firestore/lite';
-import { db } from '../firebase/firebase_config';
+import { db } from '../../firebase/firebase_config';
 
 export default async function postCustomer(
   schoolEmail,
